APG-273 Moving person methods out of referral service
