+++ src/test/resources/simulations/__files/prison-search-results-A1234AA.json
@@ -1,0 +1,51 @@
+[
+  {
+    "prisonerNumber": "A1234AA",
+    "bookingId": "1201102",
+    "bookNumber": "38518A",
+    "firstName": "JOHN",
+    "lastName": "SMITH",
+    "dateOfBirth": "1940-12-20",
+    "gender": "Male",
+    "youthOffender": false,
+    "status": "INACTIVE TRN",
+    "lastMovementTypeCode": "TRN",
+    "lastMovementReasonCode": "MED",
+    "inOutStatus": "TRN",
+    "prisonId": "TRN",
+    "prisonName": "Transfer",
+    "aliases": [
+      {
+        "firstName": "DOUGLAS",
+        "lastName": "ADORNO",
+        "dateOfBirth": "1939-11-19",
+        "gender": "Male",
+        "ethnicity": "Asian/Asian British: Indian"
+      }
+    ],
+    "alerts": [
+      {
+        "alertType": "H",
+        "alertCode": "HA2",
+        "active": true,
+        "expired": false
+      }
+    ],
+    "legalStatus": "RECALL",
+    "imprisonmentStatus": "CUR_ORA",
+    "imprisonmentStatusDescription": "ORA Recalled from Curfew Conditions",
+    "recall": true,
+    "indeterminateSentence": false,
+    "receptionDate": "2021-01-08",
+    "locationDescription": "Transfer",
+    "restrictedPatient": false,
+    "currentIncentive": {
+      "level": {
+        "code": "ENH",
+        "description": "Enhanced"
+      },
+      "dateTime": "2022-06-08T14:41:18",
+      "nextReviewDate": "2023-06-08"
+    }
+  }
+]
