+++ src/test/resources/simulations/mappings/prison-search-api.json
@@ -21,6 +21,22 @@
         "urlPattern": "/prisoner-search/prisoner-numbers",
         "method": "POST",
         "bodyPatterns" : [ {
+          "contains" : "A1234AA"
+        } ]
+      },
+      "response": {
+        "status": 200,
+        "headers": {
+          "Content-Type": "application/json"
+        },
+        "bodyFileName": "prison-search-results-A1234AA.json"
+      }
+    },
+    {
+      "request": {
+        "urlPattern": "/prisoner-search/prisoner-numbers",
+        "method": "POST",
+        "bodyPatterns" : [ {
           "contains" : "C6666DD"
         } ]
       },
